### 2023-10-13 #5945 @gchure
Hi!  Below is a short list of reviewers I would recommend for this manuscript. I don't know any of them personally, but I think they would have the domain expertise and the technical acumen to give `hplc-py` a fair shake:

* Florian Huber (https://reviewers.joss.theoj.org/reviewers/1274)
* Adriano Rutz (https://reviewers.joss.theoj.org/reviewers/2323)
* Adam Krajewski (https://reviewers.joss.theoj.org/reviewers/3189)
