### 2023-10-15 #5945 @Kevin-Mattheus-Moerman
@gchure thanks for this submission. I am the AEiC on this track, and here to help process the initial steps. Note that I've just triggered a scope review by the editorial board. This is mainly because of the relatively small size (e.g. in terms of number lines of code) of this submission. The scope review will aim to determine if this worked is deemed to conform the our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The scope review usually takes about 2 weeks to complete.
