### 2023-10-16 #5945 @gchure
Thanks @Kevin-Mattheus-Moerman. I  just want to confirm that the below software report was initiated on the `main` branch and not on `publication`. I separated the paper software and actual scientific software so I could more clearly provide the scripts used to make the figure in the paper.

My own codecount (using [VSCodeCounter](https://marketplace.visualstudio.com/items?itemName=uctakeoff.vscode-counter)) on the `hplc-py` directory lists 748 lines of code ( which I know is still below the 1000 lines of code under the scholarly effort criteria).

> ```
> Software report:
>
> github.com/AlDanial/cloc v 1.88  T=0.02 s (536.9 files/s, 62577.5 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> Python                           4             80             87            415
> Markdown                         3             26              0            223
> TeX                              1             13              0            187
> YAML                             1              0              0             18
> -------------------------------------------------------------------------------
> SUM:                             9            119             87            843
> -------------------------------------------------------------------------------
>
>
> gitinspector failed to run statistical information for the repository
> ```
