### 2023-10-16 #5945 @gchure
Here's the report generated on my end on `main/hplc`:

```
Date : 2023-10-16 10:19:27
Directory : hplc-py/hplc
Total : 3 files,  748 codes, 576 comments, 98 blanks, all 1422 lines

Languages
+----------+------------+------------+------------+------------+------------+
| language | files      | code       | comment    | blank      | total      |
+----------+------------+------------+------------+------------+------------+
| Python   |          3 |        748 |        576 |         98 |      1,422 |
+----------+------------+------------+------------+------------+------------+

Directories
+--------------+------------+------------+------------+------------+------------+
| path       | files      | code       | comment    | blank      | total      |
+--------------+------------+------------+------------+------------+------------+
| .            |          3 |        748 |        576 |         98 |      1,422 |
+--------------+------------+------------+------------+------------+------------+

Files
+------------------+----------+------------+------------+------------+------------+
| filename         | language | code       | comment    | blank      | total      |
+------------------+----------+------------+------------+------------+------------+
| hplc/__init__.py | Python   |          5 |          0 |          1 |          6 |
| hplc/io.py       | Python   |         35 |         23 |          5 |         63 |
| hplc/quant.py    | Python   |        708 |        553 |         92 |      1,353 |
| Total            |          |        748 |        576 |         98 |      1,422 |
+------------------+----------+------------+------------+------------+------------+
```
