### 2023-10-20 #5945 @Kevin-Mattheus-Moerman
I repeated `cloc` on the main branch and got the following, which is in agreement with the authors count too:

```

Language                     files          blank        comment           code
CSV                             30              0              0         768164
SVG                              1              3              1           5601
Python                          12            284            710           1444
Jupyter Notebook                11              0           5022            368
YAML                             4              4              0             86
reStructuredText                 4             32             63             47
Markdown                         1             12              0             34
DOS Batch                        1              8              1             26
CSS                              1              2              0             11
make                             1              4              7              9
SUM:                            66            349           5804         775790
```
