### 2023-11-06 #5945 @gchure
Hi @Kevin-Mattheus-Moerman, hope you're doing well! I just wanted to check on the status of this scope query. If there is any other information I can provide to help, please let me know!
