### 2023-11-21 #5945 @Kevin-Mattheus-Moerman
@gchure apologies for the delay. I have just removed the `query-scope` label as some editors felt this work may be in scope for JOSS. We will now proceed to find a handling editor.
