### 2024-01-23 5bbef11
Fixes #10 and redesigns test data for many peak filtering to be more representative of real data.
