+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 import pathlib
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 
 # The directory containing this file
 HERE = pathlib.Path(__file__).parent
