### 2024-02-06 3139fb2
Added documentaion details and community guidelines to the README
