### 2024-02-06 #6270 @gchure
Hi @florian-huber, thanks for being game to review! Here are quick responses to your comments:

1. `hplc-py` has [extensive online documentation](https://cremerlab.github.io/hplc-py) with a [quickstart](https://cremerlab.github.io/hplc-py/tutorials/quickstart.html), [example calibration curve](https://cremerlab.github.io/hplc-py/tutorials/calibration_curve.html), [step-by-step details of the algorithm](https://cremerlab.github.io/hplc-py/methodology/problem.html), and [API documentation](https://cremerlab.github.io/hplc-py/quant.html). I've now made this more clear on the repo `README`. I've also added a [mini-version](https://github.com/cremerlab/hplc-py?tab=readme-ov-file#usage) of how to use `hplc-py` (including loading, viewing, and fitting chromatograms) to the `README` so a user can really quickly see what the package is.

2. I think the [step-by-step explanation of the algorithm](https://cremerlab.github.io/hplc-py/methodology/problem.html) should satisfy more developer information, but I will include a short diagram into figure 2 of the paper. You are correct that the meat of `hplc-py` is in the `Chromatogram` class and its methods. Thanks for the suggestion!

3. I've added a [short description](https://github.com/cremerlab/hplc-py?tab=readme-ov-file#contributing) of how to contribute to `hplc-py` on the repo `README`. Please let me know if you would like more detail than is provided there. Given the anticipated use of the software, I think contribution through the issue tracker will be manageable for the maintainers (primarily me) to triage. If it somehow becomes more popular, this can be revised with more guidelines for forking and submitting pull requests.

Please let me know if you need any more info!
