### 2024-02-08 #6270 @Kastakin
I am sorry for the late review!

I read the article and went through the use of the code thanks to the well written "Quickstart" and "How it works" sections.

I do not have any particular remarks regarding, just some minors nitpicks regarding the article:

- In Figure 1 the text and code in the bottm left quadrant have significantly lover DPIs than the rest of the figure. It could be nice if it was possible to have an homogeneous figure.
- Line 64: Equation 1 is not clearly annotated in the article.

Other than that I do not see any particular problem with the library, which I enjoyed very much!
