### 2024-02-08 #6270 @gchure
Thanks @Kastakin for the feedback! I've addressed your comments in the regenerated PDF. @florian-huber , I've also included a schematic of the hplc-py data model in Figure 2, now panel (A), and addressed it in lines 36 - 41 of the new draft. Please let me know if this addresses your concern about developer documentation.
