### 2024-02-09 #6270 @florian-huber
Thanks for nice additions to the readme @gchure ! That looks all good to me now.
The given install instructions did not work straight away, because the dependencies are currently only in the requirements.txt. I suggested to add them to setup.py --> https://github.com/cremerlab/hplc-py/pull/13
