### 2024-02-09 #6270 @florian-huber
I went through it again. Looks very good, only two minor points remain from my side:
- make sure dependencies are part of pip install process (see comment above)
- Link to code for Fig. 3 is not correct (is it `Fig3_overlap.py`?)
