### 2024-02-09 3ea6285
Updating workflow by suggestion of @florian-huber. This will break the build but should get patched when I merge #13
