+++ .github/workflows/pytest.yaml
@@ -24,7 +24,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pytest-cov
-        pip install -r requirements.txt
         pip install -e ./
     - name: Test with pytest
       run: |
