### 2024-02-09 a95abe0
Fixed link to Figure 3 script, as identified by @florian-huber
