+++ CITATION.cff
@@ -7,8 +7,8 @@ authors:
 - family-names: "Cremer"
   given-names: "Jonas"
   orcid: "https://orcid.org/0000-0003-2328-5152"
-title: "hplc-py v0.1.01"
-version: 0.2.0
+title: "hplc-py v0.2.4"
+version: 0.2.4
 doi: 10.5281/zenodo.8197910
 date-released: 2023-07-30
 url: "https://github.com/cremerlab/hplc-py"
