### 2024-02-09 #6270 @gchure
Hi @florian-huber. Thanks for the PR! I've [merged it in](https://github.com/cremerlab/hplc-py/commit/ddcdf756f6934ebd0edf251c4fac4a5ac0cb0a720) and redeployed on PyPi ([0.2.4](https://pypi.org/project/hplc-py/)). This should take care of the installation issue @jgostick. Also, I've have now fixed the link to Figure 3, which points to the right script.

Thanks to both @Kastakin and @florian-huber for a quick review while paying attention to the details. This is my first review on JOSS and boy, does it make me wish this type of review process could be implemented everywhere.

@jgostick Please let me know if there is anything else you need on my end to close out the review.
