### 2024-02-09 #6270 @jgostick
@gchure You have a few more jobs to do...can you look at the above bullets about minting a DOI, assinging a final version, etc?
