### 2024-02-09 #6270 @gchure
I've cut the reviewed release of hplc-py as 0.2.5 ([PyPi record](https://pypi.org/project/hplc-py/)). It is archived on Zenodo with the doi [10.5281/zenodo.10642439](https://doi.org/10.5281/zenodo.10642439). I've confirmed that the author affiliations as well as the ORCID's are correct and match that of the Zenodo archive. The software is released under the standard GPL v3.0 license, and is specified on the Zenodo archive, the GitHub repo, and the PyPi record.
