### 2024-02-10 #6270 @jgostick
Hi @gchure
There are a few issues to tidy up with the paper:
1. The Moosavi, S. M., Ghassabian paper has their names twice
2. The capitalization on the paper title is not consistent.  I don't know if lower case or capital case is preferred, but consistency is key
3. On line 36, "Figure 2 (A)" should be "Figure 2(A)"
4. The bibliographic entry for "Choikhet, K., Glatz, B., & Rozing, G." seems incomplete
