### 2024-02-10 1e50416
Tweaks to paper given @jgostic's comments.
