+++ paper.bib
@@ -16,8 +16,7 @@
   year       = {2018},
   note       = {Publisher: Public Library of Science},
   keywords   = {Computer software, Software tools, Source code, Glycosylation, Antibodies, Background signal noise, Data processing, High performance liquid chromatography},
-  pages      = {e0200280},
-  file       = {Full Text PDF:/Users/gchure/Zotero/storage/J4XDK3CX/Jansen et al. - 2018 - HappyTools A software for high-throughput HPLC da.pdf:application/pdf}
+  pages      = {e0200280}
 }
 
 
@@ -56,11 +55,16 @@
 }
 
 @article{choikhet2003,
-  title    = {The {Physicochemical} {Causes} of {Baseline} {Disturbances} in {HPLC},},
+  series   = {{LCGC} International},
+  volume   = {6},
+  number   = {2},
+  title    = {The {Physicochemical} {Causes} of {Baseline} {Disturbances} in {HPLC}: TFA-Containing Eluents},
+  journal  = {LCGC International},
   language = {en},
   author   = {Choikhet, K and Glatz, B and Rozing, G},
   year     = {2003},
-  file     = {Choikhet et al. - 2003 - The Physicochemical Causes of Baseline Disturbance.pdf:/Users/gchure/Zotero/storage/2NPU5D5P/Choikhet et al. - 2003 - The Physicochemical Causes of Baseline Disturbance.pdf:application/pdf}
+  pages    = {96--105},
+  url      = {https://api.semanticscholar.org/CorpusID:19173011}
 }
 
 @article{felinger2004,
@@ -78,8 +82,7 @@
   month    = jul,
   year     = {2004},
   keywords = {Baseline noise, HPLC, Wavelet analysis},
-  pages    = {225--232},
-  file     = {ScienceDirect Snapshot:/Users/gchure/Zotero/storage/LFWVA4QJ/S0169743904000395.html:text/html}
+  pages    = {225--232}
 }
 
 @article{mecozzi2014,
@@ -96,8 +99,7 @@
   month    = jan,
   year     = {2014},
   keywords = {Baseline Drift Correction, Chromatographic Analysis, Environmental Monitoring, Signal Processing},
-  pages    = {2--6},
-  file     = {ScienceDirect Full Text PDF:/Users/gchure/Zotero/storage/85HKW39B/Mecozzi - 2014 - A Polynomial Curve Fitting Method for Baseline Dri.pdf:application/pdf;ScienceDirect Snapshot:/Users/gchure/Zotero/storage/D34GTGP7/S2212670814001535.html:text/html}
+  pages    = {2--6}
 }
 
 @article{macko2001,
@@ -116,8 +118,7 @@
   year       = {2001},
   note       = {Publisher: Taylor \& Francis
                 \_eprint: https://doi.org/10.1081/JLC-100103447},
-  pages      = {1275--1293},
-  file       = {Full Text PDF:/Users/gchure/Zotero/storage/DEEZ7UAF/Macko and Berek - 2001 - Pressure Effects in Hplc Influence of Pressure an.pdf:application/pdf}
+  pages      = {1275--1293}
 }
 
 @article{morhac2008,
@@ -135,8 +136,7 @@
   month    = jan,
   year     = {2008},
   note     = {Publisher: SAGE Publications Ltd STM},
-  pages    = {91--106},
-  file     = {SAGE PDF Full Text:/Users/gchure/Zotero/storage/EQPR7WCJ/Morháč and Matoušek - 2008 - Peak Clipping Algorithms for Background Estimation.pdf:application/pdf}
+  pages    = {91--106}
 }
 
 @article{choi2017,
@@ -171,8 +171,7 @@
   month    = feb,
[…]
