### 2024-02-10 #6270 @gchure
Done! I chose to keep `hplc-py` in lower case, but have corrected capitalization for the rest of the title. I also replaced "for the rapid" with "for rapid" as it reads more easily. The references have now also been fixed.
