### 2024-02-10 51a3674
Fixing capitalization in paper titles for consistency
