+++ paper.bib
@@ -1,6 +1,6 @@
 
 @article{jansen2018,
-  title      = {{HappyTools}: {A} software for high-throughput {HPLC} data processing and quantitation},
+  title      = {{HappyTools}: {A} {Software} for {High}-{Throughput} {HPLC} {Data} {Processing} and {Quantitation}},
   volume     = {13},
   issn       = {1932-6203},
   shorttitle = {{HappyTools}},
@@ -69,7 +69,7 @@
 
 @article{felinger2004,
   series   = {Advances in {Chromatography} and {Electrophoresis} - {Conferentia} {Chemometrica} 2003, {Budapest}},
-  title    = {Wavelet analysis of the baseline noise in {HPLC}},
+  title    = {Wavelet {Analysis} of the {Baseline} {Noise} in {HPLC}},
   volume   = {72},
   issn     = {0169-7439},
   url      = {https://www.sciencedirect.com/science/article/pii/S0169743904000395},
@@ -140,7 +140,7 @@
 }
 
 @article{choi2017,
-  title    = {Topographic prominence discriminator for the detection of short-latency spikes of retinal ganglion cells},
+  title    = {Topographic {Prominence} {Discriminator} for the {Detection} of {Short}-{Latency} {Spikes} of {Retinal} {Ganglion} {Cells}},
   volume   = {14},
   issn     = {1741-2552},
   url      = {https://dx.doi.org/10.1088/1741-2552/aa5646},
