### 2024-02-10 #6270 @gchure
Ah, got it. I've now fixed that so they are all capitalized.
