### 2024-02-10 #6270 @jgostick
Hmmm, odd that it only finds these issues now.  I'm not even sure what those errors mean TBH.
