### 2024-02-10 1d95746
Type consistency of affiliations on paper markdown.
