### 2024-02-10 #6270 @gchure
I have a guess -- I had mixed str and int for the affiliations in the paper YAML. Just pushed a change that I think will fix it.
