### 2024-02-10 #6270 @danielskatz
@gchure - you can't use a non-integer such as "*" as an affiliation key
