### 2024-02-10 #6270 @gchure
Ah, got it. Okay, I'll flip that to 2. If I have two affiliations, can I still supply `affiliation: 1,2` in the YAML, or does that get cast to a string?
