### 2024-02-10 #6270 @danielskatz
I'm not sure - see https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for an example
