### 2024-02-10 #6270 @gchure
Alright, this should hit the spare. I removed backticks around my email for correspondence, which is what I bet threw the monospace error. Let's hit it one more time!
