### 2024-02-12 #6270 @Kevin-Mattheus-Moerman
@gchure As AEiC I will now help to process final steps for acceptance in JOSS. I have just checked your repository, this review, the paper, and the archive link. Most seems in order. However the below points require your attention:

- [ ] In your affiliations, please spell out `USA` as `United States of America`
- [ ] Update the title in your archive to match the paper title.
